pkg/fileutil: clean up interface, comments
