+++ pkg/fileutil/lock_unix.go
@@ -26,14 +26,6 @@ var (
 	ErrLocked = errors.New("file already locked")
 )
 
-type Lock interface {
-	Name() string
-	TryLock() error
-	Lock() error
-	Unlock() error
-	Destroy() error
-}
-
 type lock struct {
 	fd   int
 	file *os.File
@@ -43,7 +35,6 @@ func (l *lock) Name() string {
 	return l.file.Name()
 }
 
-// TryLock acquires exclusivity on the lock without blocking
 func (l *lock) TryLock() error {
 	err := syscall.Flock(l.fd, syscall.LOCK_EX|syscall.LOCK_NB)
 	if err != nil && err == syscall.EWOULDBLOCK {
@@ -52,12 +43,10 @@ func (l *lock) TryLock() error {
 	return err
 }
 
-// Lock acquires exclusivity on the lock without blocking
 func (l *lock) Lock() error {
 	return syscall.Flock(l.fd, syscall.LOCK_EX)
 }
 
-// Unlock unlocks the lock
 func (l *lock) Unlock() error {
 	return syscall.Flock(l.fd, syscall.LOCK_UN)
 }
