+++ pkg/fileutil/lock_windows.go
@@ -25,14 +25,6 @@ var (
 	ErrLocked = errors.New("file already locked")
 )
 
-type Lock interface {
-	Name() string
-	TryLock() error
-	Lock() error
-	Unlock() error
-	Destroy() error
-}
-
 type lock struct {
 	fd   int
 	file *os.File
@@ -42,17 +34,14 @@ func (l *lock) Name() string {
 	return l.file.Name()
 }
 
-// TryLock acquires exclusivity on the lock without blocking
 func (l *lock) TryLock() error {
 	return nil
 }
 
-// Lock acquires exclusivity on the lock without blocking
 func (l *lock) Lock() error {
 	return nil
 }
 
-// Unlock unlocks the lock
 func (l *lock) Unlock() error {
 	return nil
 }
